+++ builder/infer_bounds.h
@@ -3,7 +3,7 @@
 
 #include <vector>
 
-#include "runtime/expr.h"
+#include "runtime/stmt.h"
 
 namespace slinky {
