+++ builder/node_mutator.h
@@ -2,6 +2,7 @@
 #define SLINKY_BUILDER_NODE_MUTATOR_H
 
 #include "runtime/expr.h"
+#include "runtime/stmt.h"
 
 namespace slinky {
