+++ builder/optimizations.h
@@ -1,7 +1,7 @@
 #ifndef SLINKY_BUILDER_OPTIMIZATIONS_H
 #define SLINKY_BUILDER_OPTIMIZATIONS_H
 
-#include "runtime/expr.h"
+#include "runtime/stmt.h"
 
 namespace slinky {
