+++ runtime/BUILD
@@ -10,7 +10,7 @@ cc_library(
         "buffer.cc",
         "depends_on.cc",
         "evaluate.cc",
-        "expr.cc",
+        "expr_stmt.cc",
         "pipeline.cc",
         "print.cc",
     ],
@@ -21,6 +21,7 @@ cc_library(
         "expr.h",
         "pipeline.h",
         "print.h",
+        "stmt.h",
         "util.h", 
     ],
     visibility = ["//visibility:public"],
