+++ runtime/evaluate.cc
@@ -16,6 +16,7 @@
 #include "runtime/depends_on.h"
 #include "runtime/expr.h"
 #include "runtime/print.h"
+#include "runtime/stmt.h"
 #include "runtime/util.h"
 
 namespace slinky {
