+++ runtime/evaluate.h
@@ -2,6 +2,7 @@
 #define SLINKY_RUNTIME_EVALUATE_H
 
 #include "runtime/expr.h"
+#include "runtime/stmt.h"
 
 namespace slinky {
